+++ src/main/java/lazyae2/core/ModGuiHandler.java
@@ -18,12 +18,15 @@ import net.minecraftforge.fml.common.network.IGuiHandler;
 import lazyae2.block.BlockMachine;
 import lazyae2.client.gui.GuiAggregator;
 import lazyae2.client.gui.GuiCentrifuge;
+import lazyae2.client.gui.GuiEnergizer;
 import lazyae2.client.gui.GuiEtcher;
 import lazyae2.container.ContainerAggregator;
 import lazyae2.container.ContainerCentrifuge;
+import lazyae2.container.ContainerEnergizer;
 import lazyae2.container.ContainerEtcher;
 import lazyae2.tile.TileAggregator;
 import lazyae2.tile.TileCentrifuge;
+import lazyae2.tile.TileEnergizer;
 import lazyae2.tile.TileEtcher;
 import appeng.container.AEBaseContainer;
 import appeng.container.ContainerOpenContext;
@@ -48,6 +51,9 @@ public final class ModGuiHandler implements IGuiHandler {
         if (BlockMachine.Type.of(id) == BlockMachine.Type.ETCHER && tile instanceof TileEtcher) {
             return withContext(new ContainerEtcher(player.inventory, (TileEtcher) tile), world, x, y, z);
         }
+        if (BlockMachine.Type.of(id) == BlockMachine.Type.ENERGIZER && tile instanceof TileEnergizer) {
+            return withContext(new ContainerEnergizer(player.inventory, (TileEnergizer) tile), world, x, y, z);
+        }
         return null;
     }
 
@@ -67,6 +73,10 @@ public final class ModGuiHandler implements IGuiHandler {
         if (BlockMachine.Type.of(id) == BlockMachine.Type.ETCHER && tile instanceof TileEtcher) {
             return new GuiEtcher(
                     withContext(new ContainerEtcher(player.inventory, (TileEtcher) tile), world, x, y, z));
+        }
+        if (BlockMachine.Type.of(id) == BlockMachine.Type.ENERGIZER && tile instanceof TileEnergizer) {
+            return new GuiEnergizer(
+                    withContext(new ContainerEnergizer(player.inventory, (TileEnergizer) tile), world, x, y, z));
         }
         return null;
     }